API Post and experiement with soft assertion
