+++ requirements.txt
@@ -13,6 +13,7 @@ Pygments==2.19.2
 pytest==8.4.1
 pytest-asyncio==1.0.0
 pytest-base-url==2.1.0
+pytest-check==2.5.3
 pytest-html==4.1.1
 pytest-metadata==3.1.1
 pytest-playwright==0.7.0
